changed structure of project
