+++ tetris/readme.md
@@ -1,0 +1,7 @@
+# Estrutura
+
+O projeto está separado em 3 partes atualmente
+- Core: Aqui seria teoricamente o "nucleo" caso usassemos a engine, apenas coisas essenciais
+- GUI: Aqui é a parte onde é abstraido em parte a interface visual e interação com o usuario
+- Logic: Aqui é a logica do jogo por si só
+
