+++ tetris/src/gui/shader.frag
@@ -4,5 +4,5 @@ in vec3 vertex_color;
 out vec4 color;
 
 void main() {
-    color = vec4(vertex_color, 1.0);   // We need an alpha value as well
+    color = vec4(vertex_color, 1.0);    
 }
