--- tetris/src/gui/square.rs
@@ -1,6 +1,0 @@
-use crate::vector2::Vec2;
-#[derive(Debug)]
-pub struct Rect {
-    pub size: Vec2,
-    pub center: Vec2,
-}
